Ignore links in messages from ourselves.
Don't announce the link title for messages eminating from the bot itself.

Otherwise we have the potential for loops.

For instance, in B::BB::M::P::GItHub::EasyLinks, it recognises a SHA mentioned
in the channel and announces details of that commit, along with the URL.  This
module would see the URL, and announce the page title - which includes the SHA -
which the other module would recognise and show the details for... turtles all
the way down.